orchestration/docker/requirements.txt
# Investment Signals Orchestration - Python Dependencies
#
# Core dependencies for all pipelines

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# HTTP/API clients
requests>=2.31.0
httpx>=0.25.2
aiohttp>=3.9.1

# Data processing
pandas>=2.1.3
numpy>=1.26.2

# Alerting
twilio>=8.10.2

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Scheduling
schedule>=1.2.1

# Monitoring
prometheus-client>=0.19.0

# Parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Airflow providers
apache-airflow-providers-postgres>=5.7.1
apache-airflow-providers-celery>=3.4.1
apache-airflow-providers-redis>=3.4.1
apache-airflow-providers-slack>=8.0.0
apache-airflow-providers-smtp>=1.3.2
